+++ MetroEditor/VectorLayer.m
@@ -348,7 +348,7 @@
     [[NSScanner scannerWithString:gString] scanHexInt:&g];  
     [[NSScanner scannerWithString:bString] scanHexInt:&b];  
     
-    float components[4];
+    CGFloat components[4];
     components[0] = (float) r / 255.0f;
     components[1] = (float) g / 255.0f;
     components[2] = (float) b / 255.0f;
@@ -379,7 +379,7 @@
     g = M - (M-g)*sd;
     b = M - (M-b)*sd;
     
-    float components[4];
+    CGFloat components[4];
     components[0] = r;
     components[1] = g;
     components[2] = b;
